isize -> i32

+++ src/aoc2022/day23.rs
@@ -62,14 +62,14 @@ impl Solution<Day23> for Aoc2022 {
     }
 }
 
-const DELTAS: [[(isize, isize); 3]; 4] = [
+const DELTAS: [[(i32, i32); 3]; 4] = [
     [(-1, -1), (0, -1), (1, -1)], // N
     [(-1, 1), (0, 1), (1, 1)],    // S
     [(-1, -1), (-1, 0), (-1, 1)], // W
     [(1, -1), (1, 0), (1, 1)],    // E
 ];
 
-fn next_state(state: &HashSet<(isize, isize)>, start_di: usize) -> HashSet<(isize, isize)> {
+fn next_state(state: &HashSet<(i32, i32)>, start_di: usize) -> HashSet<(i32, i32)> {
     let mut next_state = HashSet::with_capacity(state.len());
     for &(x, y) in state {
         let mut all_suitable = true;
@@ -120,9 +120,9 @@ fn next_state(state: &HashSet<(isize, isize)>, start_di: usize) -> HashSet<(isiz
     normalize(next_state)
 }
 
-fn normalize(state: HashSet<(isize, isize)>) -> HashSet<(isize, isize)> {
-    let mut xmin = isize::MAX;
-    let mut ymin = isize::MAX;
+fn normalize(state: HashSet<(i32, i32)>) -> HashSet<(i32, i32)> {
+    let mut xmin = i32::MAX;
+    let mut ymin = i32::MAX;
     for &(x, y) in state.iter() {
         if x < xmin {
             xmin = x;
@@ -138,8 +138,8 @@ fn normalize(state: HashSet<(isize, isize)>) -> HashSet<(isize, isize)> {
         .collect()
 }
 
-fn score(grid: &HashSet<(isize, isize)>) -> usize {
-    fn len(mm: MinMaxResult<isize>) -> isize {
+fn score(grid: &HashSet<(i32, i32)>) -> usize {
+    fn len(mm: MinMaxResult<i32>) -> i32 {
         match mm {
             itertools::MinMaxResult::NoElements => unreachable!(),
             itertools::MinMaxResult::OneElement(_) => 1,
@@ -156,8 +156,8 @@ fn score(grid: &HashSet<(isize, isize)>) -> usize {
 }
 
 #[allow(dead_code)]
-fn dbg_grid(grid: &HashSet<(isize, isize)>) {
-    fn range(mm: MinMaxResult<isize>) -> RangeInclusive<isize> {
+fn dbg_grid(grid: &HashSet<(i32, i32)>) {
+    fn range(mm: MinMaxResult<i32>) -> RangeInclusive<i32> {
         match mm {
             itertools::MinMaxResult::NoElements => unreachable!(),
             itertools::MinMaxResult::OneElement(x) => x..=x,
@@ -182,11 +182,11 @@ fn dbg_grid(grid: &HashSet<(isize, isize)>) {
     }
 }
 
-fn first_state(grid: &Grid<Cell>) -> HashSet<(isize, isize)> {
+fn first_state(grid: &Grid<Cell>) -> HashSet<(i32, i32)> {
     grid.iter()
         .filter_map(|(x, y, val)| {
             if *val == Cell::Elf {
-                Some((x as isize, y as isize))
+                Some((x as i32, y as i32))
             } else {
                 None
             }
